Allow to enable/disable several services, ports, etc. at once
The delimiter for forward_ports needed to be changed to ";" for this to make
sure that a single forward_port will not be split up automatically using ",".

Now a playbook can look like this example:

- hosts: localhost
  vars:
    firewall:
      - { service: [ 'tftp', 'ftp' ],
          port: [ '443/tcp', '443/udp' ],
	  trust: [ 'foo', 'bar' ],
	  masq: [ 'foo2', 'bar2' ],
	  forward_port: [ 'eth0;447/tcp;;1.2.3.4', 'eth0;448/tcp;;1.2.3.5' ],
          state: 'enabled' }
      - { trust_by_mac: "00:11:22:33:44:55", state: 'enabled' }
      - { masq_by_mac: [ "11:22:33:44:55:66", "11:22:33:44:55:67", ],
          state: 'enabled' }
      - { forward_port_by_mac: '00:11:22:33:44:55;447/tcp;;1.2.3.4',
          state: 'enabled' }
  roles:
    - ansible-role-firewall

The _by_mac tasks are separate to make sure that the others are succeeding.
If there in an error in the block, the whole block will not get applied.

The playbooks 1 and 2 have been updated for the forward_port delimiter. There
are new playbooks 3 and 4 that are showing the use of blocks.